// hosmas-next-14/src/types/hostels.ts
export interface hostel {
    name: string;
    path: string;
    floors: number;
    rooms: number;
    students: number;
    warden: string;
    assistantWarden: string;
    dayCaretaker: string;
    nightCaretaker: string;
    contact: string;
    emailW: string;
    emailC: string;
    gender: string;
    image: string[];
    description: string;
  }
  